shmem code now is separate source file
